Clean up least recently used entries in jars cache
The files in the jars cache are now automatically cleaned up based on a
least recently used strategy. The structure in the cache is as follows:

 - hash 1 (*)
   - some.jar
 - hash 2 (*)
   - another.jar

The cache now keeps track of files accessed inside immediate children
of the cache directory. The cleanup action then deletes all such files/
directories that have not been accessed in the last 7 days.

Issue: #1085